requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy

face_recognition
dlib

ultralytics
supervision

python-dotenv
telepot
opencv-python
Pillow
requests
numpy